fb_mainApplicationWelcomeScreen (#20)
* chore: Fix typos and add language selection label in sign up form

- Fixed typo in "languageEnglish" key in Bundle_en.properties
- Added "signUpLanguageLabel" key in Bundle_en.properties for language selection in sign up form.

* feat: Updated welcome headline label and added image to main application screen

Updated the welcome headline label style and content in the main application
screen. Added an image with a clip to the ImageView. Made changes to display
the user's name dynamically in the welcome message.

* feat: Add language label to sign up dialog

Added a language label to the sign up dialog for clarity and ease of use.

* feat: Update user display name in main application controller

Replace hard-coded user name with friendly display name from MockActiveSession
to improve user experience and personalization.