fb_TariffSelectionComponent (#43)
* feat: Add new tariff information fields and titles

Added new fields and titles for tariff information in both German and English
properties files to provide more details about each tariff. Tariff name,
description, price, unit, minimum duration, cancellation period, category, and
buttons for selection and more information were included. Also added headers for
each field.

* refactor: Update UI layout for electricity, gas, and heat pump options

Restructured the FXML files to improve the layout for the electricity,
gas, and heat pump options in the signup dialogs. Added new VBoxes and
ScrollPane elements to enhance the user experience.

* feat: Add UI components for displaying Tariff information

Added UI components to display Tariff information in a dialog pane.
Updated Tariff class to include methods for creating dialog content.
Introduced TariffDisplay class to handle displaying Tariff objects.
Implemented methods for handling Tariff selection and display.

* feat: Add electricity tariffs display and selection functionality

Added functionality to display electricity tariffs and allow users to select a tariff. The user can now choose an electricity tariff before proceeding to the next step in the sign-up process.

* feat: Add tariff selection functionality to InitialSignUpControllerStep3

Add functionality to allow users to select a tariff for gas and heating options
during the initial sign-up process. Displays available tariffs and handles
selection validation before proceeding to the next step.

* refactor: Simplify Tariff constructor initialization

Simplified the Tariff constructor by removing unnecessary empty line and curly braces.

* refactor: Change access modifier of displayTariffs method to private

Changed the access modifier of the displayTariffs method in TariffDisplay class from public to private to encapsulate the method properly within the class.

* feat: Add unit tests for TariffDisplay

Add unit tests to ensure that the correct number of rows are added to the
display when different numbers of tariffs are added. This helps improve the
reliability of the TariffDisplay component.

* feat(ui): Refactor TariffDisplay to improve dialog pane creation

Extracted the dialog pane creation logic into a separate method getDialogPane 
in TariffDisplay to improve readability and maintainability of the code. 
Now, the dialog pane is created with a GridPane layout for displaying tariff 
details in a clear and organized manner.