update-selected-tariff-and-checkbox-logic (#90)
* feat: Add functionality to hide/show select button based on electricity checkbox

Added a new method onElectricityCheckBoxClicked in InitialSignUpControllerStep2
to hide/show the select button in the tariff table based on electricity checkbox
selection. Also updated the checkboxElectricity element in the FXML file to
include the onAction attribute for the new method.

* feat: Add method to hide select button in TariffTableDisplay

Added a new method setHideSelectButton to hide the select button in
TariffTableDisplay. This method sets the selected tariff to null,
hides the select button, and updates the display.

* test: Add tests for hiding and showing select button in TariffTableDisplay

Added unit tests to check the visibility of the select button in the TariffTableDisplay based on the visibilty of the button,