tariff-selection-confirmation-pop-up-fix (#103)
* feat: Update mockServerSideErrorCounter in MockHttpClient to -4

* feat: Updated tariff selection confirmation texts and messages. Added new texts for confirmation header, checkbox, buttons, error message, and title for better user understanding and experience.

* feat: Add tariff confirmation popup dialog and controller

Added FXML file for tariff confirmation popup dialog with title, text,
checkbox, buttons, and grid pane. Also added controller for handling
checkbox selection, closing the dialog on cancel or confirm actions.

* refactor: Make createTariffLabel and createTariffGridPane static

Made the createTariffLabel and createTariffGridPane methods in TariffTableDisplay class static to improve code organization and avoid instance-specific behavior.

* feat: Add TariffSelectionConfirmationPopUp component

Add a new component TariffSelectionConfirmationPopUp to handle the confirmation
of tariff selection in the initial sign-up process. This component includes
functionality to display a dialog with the selected tariff details and a
checkbox for confirmation.

The component uses a FXMLLoader to load the FXML file for the dialog, sets up
the dialog content and title, and allows the user to confirm their tariff
selection.

This new component enhances the user experience during the sign-up process by
providing clear confirmation of the selected tariff.

* feat: Add confirmation pop-up for accepting tariff selections

Add functionality to display a confirmation pop-up when accepting tariff selections in
InitialSignUpControllerStep2 and InitialSignUpControllerStep3. This ensures that the user
explicitly confirms their tariff selection before proceeding to the next step.

* fixed width

* added error popup on no checkbox selected

---------

Co-authored-by: joscha-st <114745515+joscha-st@users.noreply.github.com>