mock_branch (#116)
* refactor: Add method to handle mocking issues in MockHttpClient

Added a new method `doNotMockIssueOnPaul` to handle issues related to mocking in the `MockHttpClient` class. This method sets a flag `doMockIssues` to true when called. This method should not be used in production code to avoid mocking issues.

* refactor: Add method to handle mocking issues in MockHttpClient

Added a new method `doNotMockIssueOnPaul` to handle issues related to mocking in the `MockHttpClient` class. This method sets a flag `doMockIssues` to true when called. This method should not be used in production code to avoid mocking issues.

* refactor: Update MockHttpClient to handle issue mocking

Add a new boolean attribute, `doMockIssues`, to handle issue mocking within MockHttpClient. This change aims to improve the flexibility and customization of issue mocking functionality.

* feat: Update method call to make update billing address request

Modified InitialSignUpControllerStep1 to update the method call to make
update billing address request for user. This change ensures the correct
method is called and improves functionality.

* refactor: Remove unused method and variable in MockHttpClient and InitialSignUpControllerStep1

Removed the unused doMockIssues variable and doNotMockIssueOnPaul method in 
MockHttpClient and InitialSignUpControllerStep1 classes respectively. These 
changes were made to clean up the codebase and improve maintainability.

* feat: Refactor SignUpManager to utilize static dialogStage

Updated SignUpManager to utilize a static dialogStage variable for consistency
and ease of access in InitialSignUpControllerStep2 and InitialSignUpControllerStep3.

* refactor: Update dialogStage reference to SignUpManager in SignUpControllers.